Added a test against memory.
By default, the script will use one processor to recompile everything,
which is the safest way to compile (i.e. you are much less likely to run
out of memory).

We expect each process to use around 4Gb of memory, so we compute the
maximum number of processors as:

  processors = free-memory / 4

and compare that with the number you specified with --processors=...
and if you used a larger number, we warning you about compiling with that
many processors. (i.e. you're likely to blow your memory away and swap
like crazy and it won't be any faster than using a single processor all
the way).

So if you have 16 processors, we expect you to have at least 64Gb of free
memory (so probably 128Gb total since some is used by the system and other
tasks).