Added way to update to latest.
The check-status.sh is a script I use to verify that all the
contribs are up to date. When testing on a different system,
it is often that the contribs are a bit behind because I want
to test all of that... So this script helps me be up to date
with all the latest.