Added code to auto-update remote status.
This version finally includes the necessary code to get the current status
of our packages from launchpad. It will automatically update the status
in the table with the new status whenever it changes.

What is still missing is a color change to clearly see what is building,
what needs to be built, what was built successfully, and what failed.
There may be a few other useful states (i.e. a project that was never
built on launchpad), but the main ones are the most useful ones.