Added the necessary code to auto-update remote.
When a build is started, we mark that specific project as being in need of
an update. These updates will happen once per minute. It will reload the
PPA status file and analyze it. Once the packages were created for all
the versions, the project is marked as built and the checks stop. The
checks also stop when an error is detected.

As a result, we also show the rows in our table with a color defining the
current build state (success, failure, building, ready for next build...)