Added two new projects, removing log4cplus.
The log4cplus project won't be used anymore. It's dead again and although
it _works_ it does not offer configuration files that make it easy to setup
the logger. Also there was always an issue with threads. See snaplogger as
the replacement.

The libmimemail is a copy of the email relater functions from the
libsnapwebsites which we want to have access to from the sitter project.

The serverplugins is a copy of the plugin classes found in the cppthread
(which originally was in the libsnapwebsites library). This version works
great with any project instead of just the snapwebsites server (and the
snapwatchdog and snapmanager to some extend). This will allow us to add
plugins all over the place (we already have a network plugin for the
snaplogger, for example).