Latest with cleaned up cmake files.
I went through the list of cmake files in our contribs to clean them up
(nearly all of them).

I also moved the old raise_flag tool to the archive. The new one is in the
communicatord project where the flags are implemented in that library. The
tool will be able to read and write to the flag files using a seteuid() call.