Removed two old settings from .gitmodules.
The ipmgr is now properly listening for triggers.
Worked a bit on the fluid-settings so it moves forward. Mainly, I'm looking
at integreting the communicatord class instead of having a messenger in each
process... we want that class to handle most of the basic communication for
us (i.e. main the registration and de-registration)