Latest with a better working CLI.
This CLI has the code we want in the fluid-settings library. But especially,
it now understands the STATUS message from the communicator allow us to
register with the fluid-settings daemon whenever it is up. (i.e. the previous
version would lose the connection completely if the daemon went down and back
up...). This is in link with the --watch command line option of the CLI.