Fixed the cmake-3.5 folder issue.
In as2js/tests/CMakeLists.txt, I have one include() which was failing. This
was due to the fact that we had an improper cmake-3.5 directory name in a
couple of locations. I replaced that with the path in $(CMAKE_ROOT). Also,
the SnapBuild script will add the correct CMAKE_MODULE_PATH for all of our
projects. That way, it should make it a lot easier to have include() and
other such instructions.

Updated the as2js so the tests are live again (well... only the abstract
classes and the version test at the moment, but I have hopes to get others
in gear soon, especially the JSON and dependencies such as the nodes and
lexer).