Added better FORWARD support.
Especially, the iplock now properly supports using the same chains for
different tables and you do not get the same data in all of them.

Made the mutex class non-copyable in the cppthread. It just doesn't make
sense to allow such (even if our class is capable of being copied).