Updated with -m recent support.
This version has the iplock -m recent support.

I changed the to_ipv[46]_string() in the libaddr to output either one of
the elements (IP, port, mask). Also the mask can be forced to an IP (for
the -m recent --mask <ip>).

Updated the other packages to match the new libaddr.

Added a new template in snapdev to handle a growing buffer which starts as
a static buffer.