Moved the basic XML loader to its own project.
The prinbee database loads its configuration information from .xml files.
It uses a very basic load which only supports basic tags, attributes and
text within tags. It also skips comments and processor entries. This is
now in the basic-xml project. It will be converted in the prinbee to use
that new implementation. This new implementation supports istream as input
instead of just a filename.