Updated the mk script.
Made progress on the as2js compiler.
Added templates to allow for enumeration ++, --, + int, - int because I
use that a lot in unit tests. It should not be used outside of unit tests
and even then, only where you loop through a complete enumeration (i.e.
an enum with gaps probably needs a different mechanism to be checked in
full).