First version with functional code.
This version of as2js is now capable of generating binary code that can then
be executed at full speed (although not optimized, it is still faster than
having to compiler and running the code through a VM. However, much is still
to be done (i.e. the code only supports 4 operators for INTEGERs).