Many updates to use names.an for messages.
The eventdispatcher messages were all hardcoded inline. This is dangerous
since we may make a mistake with the name of the item. For example, the
communicatord was named just "communicator" in two places. We need to make
sure that the name doesn't change between each instance. Using the names.an
is the safest way to ensure that it remains valid everywhere.

I made quite a few other clean ups too.