Bumped compat version to 15.
The latest available version is defined in the debhelper man page. 15 is
great as it allows us to have one type of entries whether we hav a single
package or multiple packages in a project. I'll be able to remove some
-doc packages which are very small and not worth having.

Fixed the as2js handling of char_code_at() (the result is a pointer to an
integer, not to a binary variable structure).