Now detect whether packages are available.
The process on Launchpad is to check the data, start building, and then
publishing the resulting .deb. The JSON says "Build successful" way before
the .deb are available and there isn't a state change in that JSON to tell
us whether the .deb is now available or not.

Instead, I have a process which checks for the .deb using cURL. This checks
once every 5 min. and marks the process complete only after the the .deb is
available. Somehow, checking whether the .deb is available can take a minute
or two per distribution. I'm not too sure why it is so slow. Maybe at some
point I'll figure that out and fix the issue... but at least that shows me
the correct status in my local window.