Start/watch build process from background thread.
Made many functions thread safe by calling protected functions or using a mutex.
In some cases I added functions to make it a lot simpler.
I also check that only the background thread calls certain project functions.
Transform some debug logs using lower severity such as TRACE and DEBUG.
Changed date format so both dates look alike (Y-m-d T).
Tweaked the generation of the SVG to only use memory (instead of a file on disk).
Added support for colors in the SVG image, it reuses the same colors as in the table cell background.
Only update the project that was modified now that we have the capability (build, commit, pull, push, meld, etc.)
Moved the `--launchpad-url` default to the command line options.
Allow for running the tool in gdb (path can be a full path).
Updated the TODO file.
Started fixing the .conf file, but since we do not really install this tool... (and I think that the distribution & architecture_names are incorrect now).
Fixed reference to GPL 2 in the About window. It is now GPL 3.