Added logs to named.conf.options.
The ipmgr tool would add security logs (rarely used) but not the standard
logs to the /var/log/named folder. Now the setup happens when installing
the module.

iplock has to restart the rsyslogd and it now uses systemctl which should
continue to work for a long time (whereas, the old script may disappear at
some point).