Added the unique_number class to snapdev.
This was a class in the snapwebsites library used to generate a unique
number of any machine. At the moment, this implementation is only for the
current machine. It will be easy to add a few more features as required
later.

Removed the sub-.gitignore file from libtld since I don't generate the
tld_data.c file in there anymore.