Latest support for cluck.
This includes many changes to the eventdispatcher to support a new language
(reporter) with more features. It also includes a new
message::check_parameters() feature which verifies our message parameters
before calling the callback. This is really useful to avoid having to test
such in each and every message handler. Especially, if the message handler
was to call get_parameter() for a non-existant parameter (without first
testing with has_parameter()...) then the server would throw, which is
great in a development environment but rather terrible in a production
setup where we want to keep services up and running 24/7.