Added the new rununittests to all projects.
There are still about 3 projects where the tests do not pass. There are 2
or 3 that do not have tests. There are a few without much tests even if they
do have a unittest and several catch_... files... so there is still plenty
of work.

That being said, this will simplify moving forward as in: I'll be able to
get the build server to run the unittest command as required by each project
and that way I can make sure that it does not break over time (which it does
little by little otherwise).