Many small updates to get a new version out.
I'd like to start testing the sitter which now includes a way to test
certificates to make sure they do not time out. This means having a
version that fully compiles on launchpad and that includes the prinbee
database because the iplock depends on it (for the list of IPs that
need to be blocked on a restart of a server).