Added a --publish command line option.
This way I can publish all the results on the lcov.snapwebsites.org
website along the coverage results. That is, this also runs the
coverage tests too, but instead of doing so to get the coverage
output, we do it here to know whether the tests run properly or
not.