Fixed use of /run/user/${UID}.
Using that folder is great only for applications that we know can only
run when a user is logged in. For anything else, it's not a good location
to consider. So I changed a couple of projects that were using that
folder (one which was really improper, the other was okay for the most
part but failed in one of my setups).