Fixed the expr usage that would break run-tests
The expr command returns an exit code other than 0 when the result is not
zero (not always, though...) Thus the default -e flag stops a script as if
an error occurred. This fixes the expressions I added to compute the time
the whole process took.

Moved the git pull of the main folder back to the script. I does not
work inside check-status.sh because that one verifies that there is a
debian folder which is non-existent in the root of snapcpp.