Prevent messages from being cut from unit tests.
The snaplogger will cut messages that are longer than 1,000 characters. This
is a good idea in production where a ton of logs go to disk and use space.
It is also unlikely useful to have a complete message when more than 1,000
characters.

However, when running tests, not having the whole entire message can make
it harder to debug the issue(s). So by default, the snapcatch2 extension
found in snaplogger will not remove that limit.

Also fixed small errors in the docs.