Further clean up removing another boost reference.
The serverplugins was still using boost to handle the signals. In truth,
I created the snapdev::callback_manager a while back with the idea that
it could replace that functionality and it does. Especially since I
just added support for functions that either return void or bool.

Added tests, cleaned up documentations, got other projects to compile with
the new model.